requirements.txt
py-clob-client>=0.14.0
anthropic>=0.40.0
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0
python-dotenv>=1.0.0
rich>=13.7.0
schedule>=1.2.0
sqlalchemy>=2.0.0
datasets>=2.18.0
huggingface-hub>=0.21.0
pyarrow>=15.0.0
aiosqlite>=0.20.0
